feat(chat): implement chat message handling and UI components
- Added types for MessageType, ParticipantType, PlaceInfo, and ScheduleItem in a new index.ts file.
- Created ChatMessage component to display individual chat messages with sender info and optional place details.
- Developed LeftPanel component for map and calendar views, integrating GoogleMap and WeeklyScheduleView.
- Implemented MiddlePanel to manage liked places with filtering options and tab navigation.
- Built RightPanel for displaying chat messages and input area, including functionality for toggling confirmed and favorite places.
- Refactored chat.tsx to utilize the new panel components and manage state effectively.
- Updated WeeklyScheduleView to export ScheduleItem type for better type management.
